separate setupScenes func
