+++ packages/engine/src/constants.ts
@@ -1,0 +1,1 @@
+export const ACTIVE_SCENE_ID_NODE_ID = 'active-scene-id'
