+++ packages/engine/src/index.ts
@@ -10,3 +10,4 @@ export { findNodeWithKeyFromIdList } from './utils/findNodeWithKeyFromIdList'
 export { nodesAsArray } from './utils/nodesAsArray'
 export { addNode } from './store/shared/addNode'
 export { fastIsEqual as isEqual } from 'fast-is-equal'
+export * from './constants'
