+++ packages/engine/src/store/actionCreators/scene.ts
@@ -6,14 +6,10 @@ export const createAddScene: SetterCreator<'addScene'> = (setState) => () => {
   const newSceneId = createUniqueId()
 
   setState((state) => {
-    const sceneCount = Object.values(state.nodes).filter(
-      (node) => node?.nodeType === 'scene',
-    ).length
-
     state.nodes[newSceneId] = {
       id: newSceneId,
       nodeType: 'scene',
-      title: `Scene ${sceneCount + 1}`,
+      title: `Scene ${state.sceneIds.length + 1}`,
       parentIds: [],
       childGroups: {
         optionNodeIds: [],
